Update 'gnatsd' to v1.0.2.
This updates the bundled version of 'gnatsd' to v1.0.2. I've also
rebuilt 'route_checker' using Go 1.9.0 and updated 'README.md' and the
deployment descriptors accordingly.

I've also bumped to 'alpine:3.6', and replaced the 'sudo' call in
'run.sh' with 'exec', which makes 'gnatsd' PID 1, which in turn allows
for better signal handling (the container won't be forcefully killed
anymore, allowing for NATS to exit cleanly).